requirements.txt
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
pandas==2.0.2
selenium==4.11.2
pyarrow==12.0.1
db-dtypes==1.1.1